add notification->models, refactor Notification.js
